Refactor CryptoType to use better names, and size_t not uint32_t
Test: treehugger
Bug: 147814592

+++ cryptfs.h
@@ -74,7 +74,7 @@ const char* cryptfs_get_password(void);
 void cryptfs_clear_password(void);
 int cryptfs_isConvertibleToFBE(void);
 
-uint32_t cryptfs_get_keysize();
+size_t cryptfs_get_keysize();
 const char* cryptfs_get_crypto_name();
 
 #endif /* ANDROID_VOLD_CRYPTFS_H */
